grpc_service/to_do/pyproject.toml
[tool.poetry]
name = "todo"
version = "0.1.0"
description = ""
authors = ["Jeff Su <jeffdsu@gmail.com>"]
readme = "README.md"
packages = [{include = "grpc_try"}]

[tool.poetry.dependencies]
python = "^3.11"
grpcio = "^1.59.3"
grpcio-tools = "^1.59.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
